==> 01_1_prepare_test_dataset_list.ipynb <==
import os

import pandas as pd

dataset = pd.read_csv("datasets/deep-fashion/In-shop Clothes Retrieval Benchmark/Eval/list_eval_partition.txt",
                      delim_whitespace=True)
dataset = dataset[dataset["evaluation_status"] == "query"]
dataset["is_included"] = True
dataset

Remove missing image from list of images

dataset = dataset[dataset["item_id"] != "id_00007982"]

## Remove all images from test dataset that do not include a face

from deepface import DeepFace
from tqdm import tqdm
dataset['has_face'] = False
for index, row in tqdm(dataset[dataset["is_included"] == True].iterrows()):
    hires_path = row["image_name"].replace("img", "img_highres")    
    detected_faces = DeepFace.extract_faces("datasets/deep-fashion/In-shop Clothes Retrieval Benchmark/Img/" + hires_path, enforce_detection=False, align=False)
    face_detected = not all(face["confidence"] < 5 for face in detected_faces)
    dataset["has_face"][index] = face_detected
    if not face_detected:
        dataset["is_included"][index] = False
dataset

## Remove all images from test dataset that do not provide a ground truth mask

dataset["has_mask"] = False
for index, row in tqdm(dataset[dataset["is_included"] == True].iterrows()):
    hires_path = row["image_name"].replace("img", "img_highres")
    hires_path_mask = row["image_name"].replace("img", "img_highres").replace(".jpg", "_mask.png")
    mask_exists = os.path.exists("datasets/deep-fashion/In-shop Clothes Retrieval Benchmark/Img/" + hires_path_mask)
    dataset["has_mask"][index] = mask_exists
    if not mask_exists:
        dataset["is_included"][index] = False
dataset

## Equalize the number of images for MEN and WOMAN images

import numpy as np

dataset["is_man"] = False
for index, row in tqdm(dataset.iterrows()):
    is_woman = "WOMEN" in row["image_name"]
    dataset["is_man"][index] = not is_woman

assuming there are less MEN images

len_MEN = len(dataset[dataset["is_man"] == True][dataset["is_included"] == True])
len_WOMEN = len(dataset[dataset["is_man"] == False][dataset["is_included"] == True])
print("MEN", len_MEN)
print("WOMEN", len_WOMEN)

amount_WOMAN_to_ignore = len_WOMEN - len_MEN
for index, row in tqdm(dataset[dataset["is_included"] == True].iterrows()):
    if amount_WOMAN_to_ignore > 0 and not row["is_man"]:
        dataset["is_included"][index] = False
        amount_WOMAN_to_ignore -= 1
len_MEN = len(dataset[dataset["is_man"] == True][dataset["is_included"] == True])
len_WOMEN = len(dataset[dataset["is_man"] == False][dataset["is_included"] == True])
print("MEN", len_MEN)
print("WOMEN", len_WOMEN)

len_MEN = len(dataset[dataset["is_man"] == True][dataset["is_included"] == True])
len_WOMEN = len(dataset[dataset["is_man"] == False][dataset["is_included"] == True])
print("MEN", len_MEN)
print("WOMEN", len_WOMEN)

test_dataset = pd.DataFrame()
test_dataset["image_name"] = dataset[dataset["is_included"] == True]["image_name"]
test_dataset["image_name"] = test_dataset["image_name"].transform(lambda name: name.replace("img", "img_highres"))
test_dataset = test_dataset.reset_index(drop=True)
test_dataset

## Save new test dataset

test_dataset.to_csv("datasets/query_list.csv")
test_dataset
