==> 01_3_preprocess_ground_truth_segmentation_masks.ipynb <==
import os

import pandas as pd
from PIL import Image
import numpy as np
from matplotlib.image import imread
import matplotlib.pyplot as plt
from tqdm import tqdm


segmentation_colors = {
    "background": [0, 0, 0, 255],
    "pants": [211, 211, 211, 255],
    "alt_belt": [0, 100, 0, 255],
    "face": [16, 78, 139, 255],
    "alt_dress": [16, 78, 139, 255],
    "bag": [50, 205, 50, 255],
    "leggings": [70, 130, 180, 255],
    "hat": [127, 255, 212, 255],
    "alt_eyeglass": [127, 255, 212, 255],
    "hands_and_neck": [144, 238, 144, 255],
    "alt_skin": [144, 238, 144, 255],
    "alt_hair": [211, 211, 211, 255],
    "neckwear": [220, 220, 220, 255],
    "pullover": [220, 220, 220, 255],
    "shoes": [245, 222, 179, 255],
    "alt_face": [245, 222, 179, 255],
    "skirt": [250, 235, 215, 255],
    "hair": [255, 0, 0, 255],
    "alt_headwear": [255, 0, 0, 255],
    "alt_pants": [255, 140, 0, 255],
    "dress": [255, 250, 205, 255],
    "alt_outer": [255, 250, 205, 255],
    "top": [255, 250, 250, 255],
    "alt_footwear": [255, 255, 0, 255],
}

clothes_colors = np.array([segmentation_colors['pants'], segmentation_colors["pullover"], segmentation_colors["top"], segmentation_colors["shoes"], segmentation_colors["dress"]])

df = pd.read_csv("datasets/query_list.csv")
dataset_path = "/home/luca/virtual-tryon-diffusion/datasets/deep-fashion/In-shop Clothes Retrieval Benchmark/Img/"

for path in tqdm(df["image_name"]):
    segment_path = path.replace('.jpg', '_segment.png')
    if not os.path.exists(dataset_path+segment_path):
        continue
    segmentation = np.asarray(Image.open(dataset_path+segment_path))
    condensed_mask = np.zeros((segmentation.shape[0], segmentation.shape[1]), dtype=bool)
    for c in clothes_colors:
        condensed_mask = np.logical_or(condensed_mask, np.all(segmentation == c, axis=-1))
    condensed_mask = condensed_mask * 255
    Image.fromarray(condensed_mask.astype('uint8')).save(dataset_path+path.replace('.jpg', '_mask.png'))
