==> 02_1_gather_embedding_datasets.ipynb <==
Follow instructions below to create a new embedding dataset

Assuming you gathered images based on the provided manual.pdf, copy the "embedding_datasets/example_embedding_name" folder.
1. Add images to the img folder

2. use [automatic1111](https://github.com/AUTOMATIC1111/stable-diffusion-webui) webui to preprocess images, under "train" tab use "preprocess images"
3. source directory "embedding_datasets/example_embedding_name/img"
4. destination directory "embedding_datasets/example_embedding_name/img/processed"
5. width/height 1024
6. Check "Use BLIP for caption" and "Auto focal point crop"

Add "example_embedding_name, " to the start of the text captions, edit all ".txt" files in the processed folder

Run the command from the "command" file inside the [kohya-ss scripts](https://github.com/kohya-ss/sd-scripts) folder to train the embedding
