==> 03_1_face_similarity_during_training.ipynb <==
Use the "transform_test_images_for_each_embedding_training_checkpoint.py" script inside ComfyUI to generate face study images

import os
from tqdm import tqdm
from deepface import DeepFace

embedding_name = "frankfrschlng"
dataset_folder = "embedding_datasets/"+embedding_name+"/img/processed"
dataset_images_paths = []
for root, _, files in os.walk(dataset_folder):
    for f in files:
        if f.lower().endswith('.png'):
            dataset_images_paths.append(os.path.join(root, f))

transformed_images_folder = "embedding_datasets/"+embedding_name+"/results/face-study"
transformed_images_paths = []
for root, _, files in os.walk(transformed_images_folder):
    for f in files:
        if f.lower().endswith('.png') and not f.lower().startswith("mask"):
            transformed_images_paths.append(os.path.join(root, f))

with open("embedding_datasets/"+embedding_name+"/results/face_study.csv", "w") as csv:
    for transformed_image_path in tqdm(transformed_images_paths):
        avg = 0.0
        not_detected_count = 0
        for dataset_images_path in dataset_images_paths:
            result = DeepFace.verify(img1_path=transformed_image_path, img2_path=dataset_images_path,
                                     enforce_detection=False)
            if not "distance" in result.keys():
                not_detected_count += 1
            else:
                avg += result['distance']
        avg = avg / len(dataset_images_paths)
        csv.write(transformed_image_path + "," + str(avg) + "," + str(not_detected_count) + "\n", )
