==> 03_3_get_embedding_based_on_early_stopping.ipynb <==
import pandas as pd

embedding = ("frankfrschlng")
df = pd.read_csv("embedding_datasets/" + embedding + "/results/face_study.csv", sep=',', header=None)
df[0] = df[0].transform(lambda path: int(path.split("step")[1].split(".")[0]), axis=0)
df = df.sort_values(by=[0])
df = df.groupby(0).mean().reset_index()
df

patience = 3
threshold = 0.4
current = 0
best_embedding_step = None
for index, checkpoint in df.iterrows():
    if checkpoint[1] < threshold:
        current += 1
        if current == patience:
            best_embedding_step = checkpoint[0]
            break
    else:
        current = 0
else:
    best_embedding_step = df[df[1] == df[1].min()][0].iloc[0]
best_embedding_step = int(best_embedding_step.item())
print("cp embedding_datasets/"+embedding+"/model/"+embedding+"-step"+str(best_embedding_step).zfill(8)+".safetensors ~/ComfyUI/models/embeddings/"+embedding+"-step"+str(best_embedding_step).zfill(8)+".safetensors")
