==> 04_3_create_mask_overlap_plot.ipynb <==
import os
import matplotlib.pyplot as plt
import pandas as pd

embedding_names = ["fabhsr", "christphgrmmr", "frankfrschlng", "lucfrncs"]
for name in embedding_names:
    path = "embedding_datasets/" + name + "/results/transform-all/mask_overlap.csv"
    if os.path.exists(path):
        df = pd.read_csv(path, sep=',', header=None)[1]
        print(name)
        print("MEAN: ", df.mean().round(3) * 100, "%")
        print("STD: ", df.std().round(3) * 100, "%")
        print()
    else:
        print(name)
